Protect against non-existing user (and other additions)
When user did not exist or home not published yet, key was wrongly (and quietly) generated in /.ssh
Easily unnoticed because, on the next run, a new key would be generated in the right dir
Also add a new optional 'ssh_dir' parameter. This allows to create the user and generate the key during the same run, and to generate a key in a different directory (knowing it won't be exported).
Add a '$require' parameter. This allows to control dependencies, for instance, against the resource creating the user.